+++ src/main/java/org/demo/studentscore/service/ScoreService.java
@@ -1,10 +1,9 @@
 package org.demo.studentscore.service;
 
-import org.demo.studentscore.model.entity.Score;
 import org.demo.studentscore.model.vo.ScoreVO;
 
 import java.util.List;
 
 public interface ScoreService {
-    List<ScoreVO> getAllById(Long sno);
+    List<ScoreVO> getAllById(Integer sno);
 }
